+++ src/index.ts
@@ -1,6 +1,6 @@
 import * as fs from 'fs-extra';
 import * as path from 'path';
-import lockFile from 'lockfile';
+import debug from 'debug';
 import uniqid from 'uniqid';
 import Cache from 'file-system-cache';
 
@@ -42,6 +42,7 @@ export class SsgCache {
 
   public id: string;
   public cache: ReturnType<typeof Cache>;
+  public debugInstance = debug('next-ssg-cache');
   public maxTimeout = 60000;
 
   constructor() {
@@ -77,6 +78,7 @@ export class SsgCache {
         const result = await fn();
         return result;
       } catch (err) {
+        this.debugInstance('try-error: %O', err);
         if (retryCount >= maxTimes) {
           throw err;
         }
@@ -85,52 +87,12 @@ export class SsgCache {
     }
   }
 
-  public async lock(path: string) {
-    await new Promise<void>((resolve, reject) => (
-      lockFile.lock(path, (err) => {
-        if (err) reject(err);
-        resolve();
-      })
-    ));
-  }
-
-  public async unlock(path: string) {
-    await new Promise<void>((resolve, reject) => (
-      lockFile.unlock(path, (err) => {
-        if (err) reject(err);
-        resolve();
-      })
-    ));
-  }
-
-  public async waitForUnlock(path: string) {
-    const checkForLock = () => new Promise<boolean>((resolve, reject) => {
-      lockFile.check(path, (err, isLocked) => {
-        resolve(isLocked)
-      })
-    })
-
-    return Promise.race([
-      new Promise<void>(async (resolve) => {
-        while (await checkForLock()) {
-          await this.wait(100)
-        }
-      }),
-      new Promise<void>(async (resolve, reject) => {
-        await this.wait(this.maxTimeout)
-        reject(new Error(`[next-ssg-cache] Timed-out whilst waiting for cache file to unlock (${this.maxTimeout}ms)`))
-      }),
-    ]);
-  }
-
   public async get<T extends keyof SsgCacheStore>(key: T | [T, ...string[]], fetcher: () => Promise<SsgCacheStore[T]>, options?: SsgCacheGetOptions): Promise<SsgCacheStore[T]> {
     const cacheKey = typeof key === 'string' ? key : key.join('/')
 
     try {
       if (!options?.skipCache) {
-        const value = await this.retry(async () => {
-          return await this.cache.get(cacheKey, null) as SsgCacheEntry<SsgCacheStore[T]> | null
-        }, 3);
+        const value = await this.cache.get(cacheKey, null) as SsgCacheEntry<SsgCacheStore[T]> | null
 
         if (value) {
           if (
@@ -152,9 +114,6 @@ export class SsgCache {
         }
       }
 
-      await this.waitForUnlock(this.cache.path(cacheKey));
-      await this.lock(this.cache.path(cacheKey));
-
       await this.cache.set(cacheKey, {
         status: CacheStatus.PENDING,
       })
@@ -169,18 +128,14 @@ export class SsgCache {
         } : {}),
       })
 
-      await this.unlock(this.cache.path(cacheKey));
-
       return data;
     } catch (err) {
-      console.error(err);
+      this.debugInstance('get-error: %O', err);
       console.error(`[next-ssg-cache] failed to fetch/read ${cacheKey}`)
-
-      return new Promise<SsgCacheStore[T]>((resolve, reject) => {
-        setTimeout(() => {
-          this.get(key, fetcher).then(resolve).catch(reject);
-        }, 10)
-      });
+      return this.retry(
+        () => this.get(key, fetcher, options),
+        5,
+      )
     }
   }
 }