+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 ## v1.1.2
-* Implemented file locks to prevent conflicts
+* Updated retry logic
 
 ## v1.1.1
 * Fixed `undefined` bug